requirements.txt
django
graphene-django
django-graphql-jwt
